Finish Authentication and Authorization

+++ src/tasks/tasks.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { NotFoundException } from '@nestjs/common/exceptions/not-found.exception';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 
 import { CreateTaskDto } from './dto/create-task.dto';
 import { TasksFilterDto } from './dto/tasks-filter.dto';
@@ -14,31 +15,45 @@ export class TasksService {
     @InjectRepository(TaskRepository)
     private taskRepository: TaskRepository,
   ) {}
-  getTasks(tasksFilterDto: TasksFilterDto): Promise<TaskEntity[]> {
-    return this.taskRepository.getTasks(tasksFilterDto);
+  getTasks(tasksFilterDto: TasksFilterDto, user: User): Promise<TaskEntity[]> {
+    return this.taskRepository.getTasks(tasksFilterDto, user);
   }
 
-  async getTaskById(id: number): Promise<TaskEntity> {
-    const task = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<TaskEntity> {
+    const task = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!task) {
       throw new NotFoundException(`Task not found for id = ${id} `);
     }
     return task;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<TaskEntity> {
-    return await this.taskRepository.createTask(createTaskDto);
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<TaskEntity> {
+    return await this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (!result.affected) {
       throw new NotFoundException(`Task not found for id = ${id} `);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<void> {
-    const result = await this.taskRepository.update(id, { status: status });
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<void> {
+    const result = await this.taskRepository.update(
+      { id, userId: user.id },
+      {
+        status: status,
+      },
+    );
     if (!result.affected) {
       throw new NotFoundException(`Task not found for id = ${id} `);
     }
